// clients/priselci/carwow/src/lib/data/daynight-reviews.ts
export type DayNightReview = {
	id: string;
	text: string;
	avatar: `/assets/${string}`;
	name: string;
	label: string;
};

export const daynightReviews: DayNightReview[] = [];

export const daynightReviewCount = daynightReviews.length;
export const daynightReviewCountLabel = 'Няма потвърдени отзиви';
export const daynightReviewLinkLabel = `Виж всички ${daynightReviewCountLabel}`;